Extract specific massage composers.
