+++ backend/Veloci.Logic/Features/Achievements/NotificationHandlers/DiscordAchievementsHandler.cs
@@ -1,0 +1,38 @@
+using MediatR;
+using Veloci.Logic.Bot.Discord;
+using Veloci.Logic.Features.Achievements.Notifications;
+using Veloci.Logic.Features.Achievements.Services;
+
+namespace Veloci.Logic.Features.Achievements.NotificationHandlers;
+
+public class DiscordAchievementsHandler :
+    INotificationHandler<DayStreakAchievements>,
+    INotificationHandler<GotAchievements>
+{
+    private readonly DiscordAchievementMessageComposer _messageComposer;
+    private readonly IDiscordBot _discordBot;
+
+    public DiscordAchievementsHandler(DiscordAchievementMessageComposer messageComposer, IDiscordBot discordBot)
+    {
+        _messageComposer = messageComposer;
+        _discordBot = discordBot;
+    }
+
+    public async Task Handle(DayStreakAchievements notification, CancellationToken cancellationToken)
+    {
+        const int delaySec = 3;
+
+        foreach (var pilot in notification.Pilots)
+        {
+            var message = _messageComposer.DayStreakAchievement(pilot);
+            await _discordBot.SendMessageAsync(message);
+            await Task.Delay(TimeSpan.FromSeconds(delaySec), cancellationToken);
+        }
+    }
+
+    public async Task Handle(GotAchievements notification, CancellationToken cancellationToken)
+    {
+        var message = _messageComposer.AchievementList(notification.Results);
+        await _discordBot.SendMessageAsync(message);
+    }
+}
